ejercicios/semana1/booklet/sql/ejercicio_1/python_with_psycopg2/requirements.txt
psycopg2-binary
mysql-connector-python
pymysql